requirements.txt
streamlit
scikit-learn
pandas
seaborn
numpy
joblib
python-dotenv
sympy 